% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{addScrollbars}
\alias{addScrollbars}
\title{Add scrollbars to some widgets GUI}
\usage{
addScrollbars(
  parent,
  widget,
  type = c("x", "y"),
  Row = 1,
  Col = 1,
  Px = 0,
  Py = 0
)
}
\arguments{
\item{parent}{frame containing the \code{widget}}

\item{widget}{widget where to add scrollbars}

\item{type}{which scrollbars to enable}

\item{Row}{the row position of the calling widget}

\item{Col}{the column position of the calling widget}

\item{Px}{padx param. of the calling widget}

\item{Py}{pady param. of the calling widget}
}
\value{
add the selected scrollbar to the widget
}
\description{
Add scrollbars to some GUI widgets
}
