% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{asList,XPSCoreLine-method}
\alias{asList,XPSCoreLine-method}
\title{asList}
\usage{
\S4method{asList}{XPSCoreLine}(from, select = "all")
}
\arguments{
\item{from}{= an XPSCoreLine object}

\item{select}{one or few or \code{"all"} (default) of \code{c("MAIN", "RTF", "BASE", "COMPONENTS", "FIT")}}
}
\value{
'asList' returns the selected slots of a XPSCoreLine coerced in list format
}
\description{
method to coerce an object of class 'XPSCoreLine' in an object of class 'list'
}
\examples{
\dontrun{
 MyList <- asList(test[["C1s"]], selected="all")
}
}
