% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{asList}
\alias{asList}
\title{asList function for XPSCoreLine}
\usage{
asList(from, select = "all")
}
\arguments{
\item{from}{= an XPSCoreLine object}

\item{select}{one or few or \code{"all"} (default) of \code{c("MAIN", "RTF", "BASE", "COMPONENTS", "FIT")}}
}
\value{
'asList' returns the selected slots of a XPSCoreLine coerced in list format
}
\description{
asList attempts to coerce its argument to an object of class list. 
 Mainly used for plot function.
}
