% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{AsymmLorentz}
\alias{AsymmLorentz}
\title{AsymmGaussLorentz function}
\usage{
AsymmLorentz(x, h, mu, sigma, asym)
}
\arguments{
\item{x}{numeric vector}

\item{h}{function amplitude}

\item{mu}{position of function center}

\item{sigma}{function full width at half maximum}

\item{asym}{function asymmetry value}
}
\description{
Tail modified Gauss-Lorentz Sum modifed by using an exponential function
}
