% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\docType{class}
\name{baseline-class}
\alias{baseline-class}
\title{Class "baseline"}
\description{
The RxpsG package provides a class for baselines (class \code{baseline})
  \code{baseline} are values pairs stored in a \code{list}
  and additional parameters are stored in slots.
  Objects from the Class: Objects can be created by calls of the form \code{new("baseline", ...)}.
}
\examples{
\dontrun{
test <- new("baseline", ...))
}
}
