% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSCustomBaseline.r
\name{baseSpline}
\alias{baseSpline}
\title{Spline Baseline}
\usage{
baseSpline(spectra, splinePoints, limits)
}
\arguments{
\item{spectra}{coreline on which apply spline function}

\item{splinePoints}{set of points where the spline has to pass through}

\item{limits}{the abscissa range where the spline is defined}
}
\description{
definition of the Spline Algorythms
  Computes the Spline curve trough a set of splinePoints
  provided by the user by mouse-clicking on the Core-Line spectrum
  This function is called by the XPSAnalysisGUI
}
