% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{c,XPSSample-method}
\alias{c,XPSSample-method}
\title{combine}
\usage{
\S4method{c}{XPSSample}(x, ...)
}
\arguments{
\item{x}{object of class XPSSample to be combined}

\item{...}{additional parameters}
}
\value{
XPSSample
}
\description{
define a method to combine two XPSSamples
}
\examples{
\dontrun{
 XPSSampTot <- c(XPSSample1, XPSsample2)
}
}
