% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{clear_treeview}
\alias{clear_treeview}
\title{clear_treeview Clear the treeview}
\usage{
clear_treeview(widget)
}
\arguments{
\item{widget}{widget}
}
\description{
clear_treeview clears the treeview container
}
