% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{clear_widget}
\alias{clear_widget}
\title{clear_widget destroys the elements contained in a generic widget 
   like a tktoplevel(), ttkframe() or ttklabelframe()}
\usage{
clear_widget(widget)
}
\arguments{
\item{widget}{parent widget}
}
\description{
clear_widget destroys all elements in the widget container
}
