% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{ClearWidget}
\alias{ClearWidget}
\title{ClearWidget}
\usage{
ClearWidget(widget)
}
\arguments{
\item{widget}{the widget container}
}
\description{
ClearWidget removes all childrens from a widget container
}
