% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{Derivative}
\alias{Derivative}
\title{Derivative function}
\usage{
Derivative(x, mu)
}
\arguments{
\item{x}{numeric vector}

\item{mu}{position of function center}
}
\description{
Derivative function to store Derivate position in a CoreLine@Components slot
}
