% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{DFrameTable}
\alias{DFrameTable}
\title{DFrameTable}
\usage{
DFrameTable(
  Data,
  Title,
  ColNames = "",
  RowNames = "",
  Width = 10,
  Env,
  parent = NULL,
  Row = NULL,
  Column = NULL,
  Border = c(10, 10, 10, 10)
)
}
\arguments{
\item{Data}{== NAME of the object of class data.frame containing
  the items to be listed in DFrameTable if parent != NULL
Data == DATA.FRAME containing the items to be listed in 
  DFrameTable if parent == NULL}

\item{Title, }{the name of the DFrameTable object}

\item{ColNames}{the names of the DFrameTable columns}

\item{RowNames}{the names of the DFrameTable rownames}

\item{Width}{the width of the DFrameTable columns}

\item{Env}{the environment containing the object of class data.frame to print}

\item{parent}{the ID of the parent container}

\item{Row}{the row where to place the DFrameTable using tkgrid()}

\item{Column}{the column where to place the DFrameTable using tkgrid()}

\item{Border}{border width in pixels}
}
\description{
DFrameTable makes a table to edit items of a data.frame object
}
