% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{DoniachSunjic}
\alias{DoniachSunjic}
\title{DoniachSunjic function}
\usage{
DoniachSunjic(x, h, mu, sigmaDS, asym)
}
\arguments{
\item{x}{numeric vector}

\item{h}{function amplitude}

\item{mu}{position of function center}

\item{sigmaDS}{function full width at half maximum}

\item{asym}{function asymmetry value}
}
\description{
Correct version of the Doniach-Sunjic function (see Wertheim PRB 25(3), 1987, (1982))
  see Leiro et al. J. El. Spectr. Rel. Phen. 128, 205, (2003)
}
