% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSElementsTK.r
\name{ElementCheck}
\alias{ElementCheck}
\title{ElementCheck}
\usage{
ElementCheck(element)
}
\arguments{
\item{element}{character = element to check if present in the Periodic Table}
}
\description{
ElementCheck() checks if an element is present in the PeriodicTableElements
  ElementCheck is is an internal function not intended for the user
}
