% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilities.r
\name{findMaxPos}
\alias{findMaxPos}
\title{findMaxPos}
\usage{
findMaxPos(XY)
}
\arguments{
\item{XY}{numeric matrix composed by X and Y columns}
}
\value{
Returns c(x, max) values
}
\description{
findMaxPos finds the abscissa corresponding to the y-vector maximum
}
\examples{
\dontrun{
	findMaxPos(XY)
}
}
