% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilities.r
\name{findY}
\alias{findY}
\title{findY}
\usage{
findY(XY, Xvalue)
}
\arguments{
\item{XY}{numeric matrix composed by X and Y columns}

\item{Xvalue}{numeric value}
}
\value{
Returns Y value
}
\description{
findY finds the Y value corresponding to a given X value
}
\examples{
\dontrun{
	findY(XY, Xvalue)
}
}
