% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilities.r
\name{FitCompLbl}
\alias{FitCompLbl}
\title{FitCompLbl}
\usage{
FitCompLbl(position.list, label, color = "blue", ...)
}
\arguments{
\item{position.list}{list containing x and y coordinates of the label}

\item{label}{character the label the label associated to the fit-component}

\item{color}{character the color of the label}

\item{...}{additional parameters for the FitCompLbl function}
}
\description{
FitCompLbl function to draw the label identifying the fit components
  in a plotted spectrum
}
