% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitCompClass.r
\docType{class}
\name{fitComponents-class}
\alias{fitComponents-class}
\title{Class "fitComponents"}
\description{
a class that describes a fit component.  The class has the
  definition of all the necessary informations needed to use a fit component
  algorithm store all the related infromation, enable specific plot functionalities
  Objects from the Class: Objects can be created by calls of the form
  \code{new("fitComponents", ...)}.
}
\section{Slots}{

\describe{
\item{\code{funcName}}{the name of the fitting function}

\item{\code{description}}{the description of the fitting function}

\item{\code{label}}{label of the fit component associated to the added fit component}

\item{\code{param}}{the list of fitting parameters}

\item{\code{rsf}}{the RSF associated to the fit component (the same as that of the Core-Line)}

\item{\code{ycoor}}{the y values of the best fit}

\item{\code{link}}{the list containing the series of links associated to this fit component}
}}

\examples{
\dontrun{
 showClass("fitComponents")
}
}
