% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilities.r
\name{FitLin}
\alias{FitLin}
\title{FitLin}
\usage{
FitLin(X, Y)
}
\arguments{
\item{X}{abscissa}

\item{Y}{ordinate}
}
\value{
Returns c(m, c) slope m and intercept c
}
\description{
FitLin function performs a linear fit
}
\examples{
\dontrun{
	MinDist(posP0, posP1, posP2)
}
}
