% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{Generic}
\alias{Generic}
\title{Generic}
\usage{
Generic(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
a series of NA of length = length(x)
}
\description{
Generic is a function to initialize the new Fit Component Slots
}
