% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{get_column_treeview}
\alias{get_column_treeview}
\title{get_column_treeview}
\usage{
get_column_treeview(widget, column = 0)
}
\arguments{
\item{widget}{treeview widget}

\item{column}{column of the treeview widget}
}
\value{
values of selected column
}
\description{
get_column_treeview returns the indicated column of elements of the treeview GUI
}
