% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{get_common_list}
\alias{get_common_list}
\title{get_common_list Common XPS coreline for different XPS Samples}
\usage{
get_common_list(xps.sample_name)
}
\arguments{
\item{xps.sample_name}{list of xps samples name}
}
\value{
the list of XPSCoreline common to the \code{xps.sample_name}
}
\description{
get_common_list gets the list of common XPSCoreline for different XPSSamples GUI
}
\author{
Roberto Canteri \email{canteri@fbk.eu}
}
