% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{get_selected_treeview}
\alias{get_selected_treeview}
\title{get_selected_treeview The selected row of the treeview}
\usage{
get_selected_treeview(widget, column = 0)
}
\arguments{
\item{widget}{treeview widget}

\item{column}{column of the treeview widget}
}
\value{
values of selected rows and column
}
\description{
get_selected_treeview returns the selected row of the treeview GUI
}
\author{
Roberto Canteri \email{canteri@fbk.eu}
}
