% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSElementsTK.r
\name{getElementValue}
\alias{getElementValue}
\title{getElementValue}
\usage{
getElementValue(
  element,
  orbital,
  analyzer = c("scienta", "kratos"),
  what = "RSF"
)
}
\arguments{
\item{element}{character = element symbol as character}

\item{orbital}{character = atom orbital}

\item{analyzer}{character = name of reference instrument dependent table of values}

\item{what}{character = name of value to get. Default \code{"RSF"}}
}
\value{
getElementValue returns values named by \code{what}.
}
\description{
getElementValue function is used to to access and get values from the
  element libraries supplied for Scienta and Kratos instruments. This function
  is internal not intended for the user.
}
