% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{getMaxOfComponents,XPSCoreLine-method}
\alias{getMaxOfComponents,XPSCoreLine-method}
\title{getMaxOfComponents}
\usage{
\S4method{getMaxOfComponents}{XPSCoreLine}(object)
}
\arguments{
\item{object}{XPSCoreLine}
}
\description{
method to define the maximum of Fit Components
}
\examples{
\dontrun{
 print(getMaxOfComponents(test[["C1s"]]))
}
}
