% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{getMaxOfComponents}
\alias{getMaxOfComponents}
\title{getMaxOfComponents Get Max of Components}
\usage{
getMaxOfComponents(object)
}
\arguments{
\item{object}{XPSCoreLine}
}
\value{
list with \code{x,y} value
}
\description{
Get Max of Components Method for XPSCoreLine
}
