% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitCompClass.r
\name{getParam,fitComponents-method}
\alias{getParam,fitComponents-method}
\title{getParam}
\usage{
\S4method{getParam}{fitComponents}(object, parameter = NULL, variable = NULL)
}
\arguments{
\item{object}{a Core_Line object of class \code{XPSCoreLine}}

\item{parameter}{character is one of the "start", "max" or "min" parameter to be set}

\item{variable}{character as "mu", "sigma"... indicates the parameter to set}
}
\value{
'getParam' returns the value of the requested parameter
}
\description{
method to get the value of the specified fitting parameter from an object@Component[[j]]
   where object is of class 'XPSCoreLine' and j indicates the jth fit component
}
\examples{
\dontrun{
  getParam(test[["C1s"]]@Component[[2]], parameter="start", variable="mu")
}
}
