% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitCompClass.r
\name{getParam}
\alias{getParam}
\title{getParam}
\usage{
getParam(object, parameter = NULL, variable = NULL)
}
\arguments{
\item{object}{an object of class \code{XPSCoreLine}}

\item{parameter}{character is one of the "start", "max" or "min" parameter to be set}

\item{variable}{character as "mu", "sigma"... indicates the parameter to set}
}
\value{
'getParam' returns the value of the requested parameter
}
\description{
S4method 'getParam' method for get information about a fitting parameters.
}
