% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{hasBaseline,XPSCoreLine-method}
\alias{hasBaseline,XPSCoreLine-method}
\title{hasBaseline}
\usage{
\S4method{hasBaseline}{XPSCoreLine}(object)
}
\arguments{
\item{object}{a Core_Line object of class \code{XPSCoreLine}}
}
\description{
method to verify an object of class 'XPSCoreLine'
}
\examples{
\dontrun{
 if( hasBaseline(test[["C1s"]]) ) print(length(test[["C1s"]]@Baseline$y))
}
}
