% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{hasBaseline}
\alias{hasBaseline}
\title{hasBaseline}
\usage{
hasBaseline(object)
}
\arguments{
\item{object}{an object of class \code{XPSCoreLine}}
}
\value{
'hasBaseline' returns a logical vector with TRUE if the object has a Baseline FALSE otherwise
}
\description{
S4method 'hasBaseline' method for objects of class XPSCoreLine.
}
\details{
'hasBaseline' checks if Baseline is defined for an XPSCoreLine objects.
}
