% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{hasBoundaries,XPSCoreLine-method}
\alias{hasBoundaries,XPSCoreLine-method}
\title{hasBoundaries}
\usage{
\S4method{hasBoundaries}{XPSCoreLine}(object)
}
\arguments{
\item{object}{a Core_Line object of class \code{XPSCoreLine}}
}
\description{
method to verify an object of class 'XPSCoreLine'
}
\examples{
\dontrun{
 if( hasBoundaries(test[["C1s"]]) ) print(test[["C1s"]]@Boundaries)
}
}
