% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{hasBoundaries}
\alias{hasBoundaries}
\title{hasBoundaries}
\usage{
hasBoundaries(object)
}
\arguments{
\item{object}{an object of class \code{XPSCoreLine}}
}
\value{
'hasBoundaries' returns a logical vector with TRUE if the object has defined boundaries FALSE otherwise
}
\description{
S4method 'hasBoundaries' method for objects of class XPSCoreLine.
}
\details{
'hasBoundaries' checks if Boundaries are defined for XPSCoreLine objects.
}
