% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{hasComponents,XPSCoreLine-method}
\alias{hasComponents,XPSCoreLine-method}
\title{hasComponents}
\usage{
\S4method{hasComponents}{XPSCoreLine}(object)
}
\arguments{
\item{object}{a Core_Line object of class \code{XPSCoreLine}}
}
\description{
method to verify an object of class 'XPSCoreLine'
}
\examples{
\dontrun{
 if (hasComponents(test[["C1s"]])) print(length(test[["C1s"]]@Components[[1]]))
}
}
