% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{hasComponents}
\alias{hasComponents}
\title{hasComponents}
\usage{
hasComponents(object)
}
\arguments{
\item{object}{an object of class \code{XPSCoreLine}}
}
\value{
'hasComponents' returns a logical vector with TRUE if the object has defined Fit Components FALSE otherwise
}
\description{
S4method 'hasBaseline' method for objects of class XPSCoreLine.
}
\details{
'hasComponents' checks if Fit Components are defined for an XPSCoreLine objects.
}
