% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{hasFit}
\alias{hasFit}
\title{hasFit}
\usage{
hasFit(object)
}
\arguments{
\item{object}{an object of class \code{XPSCoreLine}}
}
\value{
'hasFit' returns a logical vector with TRUE if the object has a Best Fit FALSE otherwise
}
\description{
S4method 'hasFit' method for objects of class XPSCoreLine.
}
\details{
'hasFit' checks if  Best Fit is defined for an XPSCoreLine objects.
}
