% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{hasRegionToFit}
\alias{hasRegionToFit}
\title{hasRegionToFit}
\usage{
hasRegionToFit(object)
}
\arguments{
\item{object}{an object of class \code{XPSCoreLine}}
}
\value{
'hasRegionToFit' returns a logical vector with TRUE if the object has defined RegionToFit FALSE otherwise
}
\description{
S4method 'hasRegionToFit' method for objects of class XPSCoreLine.
}
\details{
'hasRegionToFit' checks if the RegionToFit is defined for XPSCoreLine objects.
}
