% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSElementsTK.r
\name{IniElement}
\alias{IniElement}
\title{IniElement}
\usage{
IniElement(element)
}
\arguments{
\item{element}{= character element to initialize}
}
\description{
IniElement() provides a table for each element of the Periodic Table.
  The table contains the element-orbitals ordered in increasing Binding Energy
  but also the relative Kinetic energy is shown together with the orbital RSF
  IniElement is is an internal function not intended for the user
  it is called by XPSSurveyElementIdentify() and XPSElemTab() functions
}
