% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{Initialize}
\alias{Initialize}
\title{Initialize}
\usage{
Initialize(x, h, mu, sigma)
}
\arguments{
\item{x}{numeric vector}

\item{h}{function amplitude}

\item{mu}{position of function center}

\item{sigma}{function full width at half maximum}
}
\description{
Initialize function to initalizes parameter slot for a generic function
}
