% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{is.notempty}
\alias{is.notempty}
\title{is.notempty}
\usage{
is.notempty(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical
}
\description{
is.notempty  is value missing, null, 0-length or NA length 1
}
