% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{plot,XPSCoreLine,missing-method}
\alias{plot,XPSCoreLine,missing-method}
\alias{missing}
\title{plot}
\usage{
\S4method{plot}{XPSCoreLine,missing}(
  x,
  type = "l",
  ltype = "solid",
  color = "black",
  main = x@Symbol,
  xlim = NULL,
  ylim = NULL,
  labels = TRUE,
  xlab = x@units[1],
  ylab = x@units[2],
  ...
)
}
\arguments{
\item{x}{The \code{XPS} numeric matrix containing a XPS CoreLine to be plotted}

\item{type}{character "l", "p", "b" for line, points or both}

\item{ltype}{character, "solid", "dashed"... pattern of the line}

\item{color}{character, "black", "red", "green"... color for the data to be plotted}

\item{main}{character, title of the plot}

\item{xlim}{numeric, range(X) limits of the X-data}

\item{ylim}{numeric, range(Y) limits of the Y-data}

\item{labels}{logical TRUE to plot axis numbers}

\item{xlab}{character "X axis label"}

\item{ylab}{character "Y axis label"}

\item{...}{additional parameters for plot function, see par()}

\item{y}{(not used) Xdata, Ydata are contained in the XPSCoreLine object}
}
\description{
method to plot objects of class XPSCoreLine or XPSSample
}
\examples{
\dontrun{
 plot(test[["C1s"]]) #plot the XPSCoreline
}
}
