% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{plot,XPSSample,missing-method}
\alias{plot,XPSSample,missing-method}
\alias{missing}
\title{'plot'}
\usage{
\S4method{plot}{XPSSample,missing}(x, y, ...)
}
\arguments{
\item{x}{XPSSample}

\item{y}{(not used) Xdata, Ydata are contained in the XPSSample object}

\item{...}{further parameters to the plot function}
}
\description{
definition of method 'plot' for objects of class 'XPSSample'
}
\examples{
\dontrun{
plot(test)
}
}
