% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{populate_treeview}
\alias{populate_treeview}
\title{populate_treeview Populate the treeview}
\usage{
populate_treeview(widget, where = "end", items)
}
\arguments{
\item{widget}{treeview widget}

\item{where}{position to add the \code{items}, default to \code{end}}

\item{items}{items to add}
}
\description{
populate_treeview populates the treeview with data saved in a \code{data.frame}
  The items MUST be a \code{data.frame}
}
\author{
Roberto Canteri \email{canteri@fbk.eu}
}
