% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSRead.Scienta.r
\name{read.scienta}
\alias{read.scienta}
\title{read.scienta function reads file data in .pxt Scienta format}
\usage{
read.scienta(file = NULL, debug = FALSE)
}
\arguments{
\item{file}{File}

\item{debug}{debug switch}
}
\description{
read.scienta function reads file data in .pxt Scienta format
}
