% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSRead.Vamas.r
\name{read.vamas}
\alias{read.vamas}
\title{read.vamas}
\usage{
read.vamas(file = NULL, debug = FALSE)
}
\arguments{
\item{file}{File}

\item{debug}{debug switch}
}
\description{
read.vamas function reads data in standard vamas format
}
