% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSRead.Genplot.r
\name{readGenplot}
\alias{readGenplot}
\title{readGenplot}
\usage{
readGenplot(Object, file = NULL, debug = FALSE)
}
\arguments{
\item{Object}{an XPSSample object where to put the read data}

\item{file}{File}

\item{debug}{debug switch}
}
\description{
Read Genplot RPL data
 Read Genplot analyzed spectra in RPL format
 a RPL folder containing analyzed data must be present
 This function is called by using the RxpsG menu
}
