% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxpsG-package.r
\docType{package}
\name{RxpsG}
\alias{RxpsG}
\title{Package RxpsG}
\description{
Processing tools for X ray Photoelectron Spectroscopy Data.
The package implements classes and methods to read, visualize and manipulate XPS spectra files.
The spectra can be obtained from .pxt and .vms data format from different instruments.'
More generally, any data that is recorded as a list of (x,y) is suitable.
}
