% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\docType{class}
\name{XPSCoreLine-class}
\alias{XPSCoreLine-class}
\title{Class "XPSCoreLine"}
\description{
The package provides a class for XPS spectra (class \code{XPSCoreLine}) and
 lists of such objects (class \code{XPSSample}). \code{XPSCoreLine} are values
 pairs stored in a \code{list} and several additional parameters stored in
 slots.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{contains the x,y spectral data}

\item{\code{RegionToFit}}{the portion of the spectrum to fit}

\item{\code{Baseline}}{the Baseline applied to subtract background}

\item{\code{Components}}{the fitting components}

\item{\code{Fit}}{the best fit}

\item{\code{Boundaries}}{the values of the RegionToFit edges}

\item{\code{RSF}}{the relative seisitivity factor associated to the element spectrum}

\item{\code{Shift}}{the energy correction shift if charging present}

\item{\code{units}}{the adopted units: kinetic/binding energy, counts/counts_per_second}

\item{\code{Flags}}{logical}

\item{\code{Info}}{information regarding the spectrum acquisition}

\item{\code{Symbol}}{symbol of the element associated to the spectrum}
}}

\examples{
\dontrun{
test <- new("XPSCoreLine", Info="test", units=c("Binding [eV]", "Counts"))
}
}
\keyword{classes}
