% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{setAsData.Frame,XPSCoreLine-method}
\alias{setAsData.Frame,XPSCoreLine-method}
\title{setAsData.Frame}
\usage{
\S4method{setAsData.Frame}{XPSCoreLine}(from, to = "data.frame")
}
\arguments{
\item{from}{= an XPSCoreLine object}

\item{to}{= a data.frame object}
}
\value{
'setAsData.Frame' returns a data.frame object
}
\description{
method to coerce an object of class 'XPSCoreLine' 
 in an object of class 'data.frame'
}
\examples{
\dontrun{
 MyDataFrame <- setAsData.Frame(test[["C1s"]], to="data.frame")
}
}
