% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{setAsMatrix}
\alias{setAsMatrix}
\title{setAsMatrix function for XPSCoreLine}
\usage{
setAsMatrix(from, to = "matrix")
}
\arguments{
\item{from}{= an XPSCoreLine object}

\item{to}{= a matrix object}
}
\value{
'setAsMatrix' returns a matrix object
}
\description{
setAsMatrix attempts to coerce an XPSCoreline to a matrix type. 
 Mainly used for plot function.
}
