% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSCustomBaseline.r
\name{Shirley0}
\alias{Shirley0}
\title{Shirley0 Classic Shirley background}
\usage{
Shirley0(object, limits)
}
\arguments{
\item{object}{XPSSample CoreLine}

\item{limits}{limits of the XPSSample object}
}
\value{
returns the Two_parameter Shirley background.
}
\description{
Shirley0 function prepresents the classic
  algorithm to define the Shirley background below a given Core-Line
  Shirley0(E) = S(E) is calculated iteratively, at step i:
  S(E)i) = integral[S(E)i-1) J(E) dE ]
  S(E)i), S(E)i-1) are the Shirley at step i and i-1
  J(E) is the measured spectrum
  This function is called by the XPSAnalysisGUI
}
