% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{SimplifiedDoniachSunjic}
\alias{SimplifiedDoniachSunjic}
\title{SimplifiedDoniachSunjic function}
\usage{
SimplifiedDoniachSunjic(x, h, mu, sigma, asym)
}
\arguments{
\item{x}{numeric vector}

\item{h}{function amplitude}

\item{mu}{position of function center}

\item{sigma}{function full width at half maximum}

\item{asym}{function asymmetry value}
}
\description{
see http://www.casaxps.com/help_manual/line_shapes.htm
}
