% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{sortComponents,XPSCoreLine-method}
\alias{sortComponents,XPSCoreLine-method}
\title{sortComponents}
\usage{
\S4method{sortComponents}{XPSCoreLine}(object)
}
\arguments{
\item{object}{a Core_Line object of class \code{XPSCoreLine}}
}
\description{
method to sort Fit Components in ascending energy (binding or kinetic)
}
\examples{
\dontrun{
 sortComponents(test[["C1s"]])
}
}
