% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{sortComponents}
\alias{sortComponents}
\title{sortComponents Sort the XPSCoreline Fit Components}
\usage{
sortComponents(object)
}
\arguments{
\item{object}{XPSCoreLine}
}
\description{
Sort Components Method for XPSCoreLine
}
