% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSCustomBaseline.r
\name{Tougaard2P}
\alias{Tougaard2P}
\title{Tougaard2P two parameter Togaard Baseline}
\usage{
Tougaard2P(object, limits)
}
\arguments{
\item{object}{XPSSample CoreLine}

\item{limits}{limits of the XPSSample CoreLine}
}
\value{
returns the 2parameter Tougaard background.
}
\description{
Tougaard2P two parameter B and C are used to define
  the Tougaard Baseline:
  Tougaard(E) = L * Integral[ K(E'-E) * J(E)dE' ]
  J(E) = acquired spectrum
  K(E'-E) = Universal Cross Section described by: B*T/(C + T^2)^2   T=E'-E
  see S. Tougaard, Surf. Sci. (1989), 216, 343;  S. Tougaard, Sol. Stat. Comm.(1987), 61(9), 547
  L = l/(1-lcos a) l=mean free path, a = take-off angle
  This function is called by XPSAnalysisGUI
}
