% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{updateTable}
\alias{updateTable}
\title{updateTable}
\usage{
updateTable(widget, items)
}
\arguments{
\item{widget}{widget to update}

\item{items}{the items used to updade the Table}
}
\description{
updateTable updates a table-like widget with the specified items
}
