% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{VBFermi}
\alias{VBFermi}
\title{VBFermi function}
\usage{
VBFermi(x, h, mu, k)
}
\arguments{
\item{x}{numeric vector}

\item{h}{function amplitude}

\item{mu}{position of function center}

\item{k}{Fermi Distribution decay rate}
}
\description{
VBFermi function
}
