% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{VBtop}
\alias{VBtop}
\title{VBtop function}
\usage{
VBtop(x, mu)
}
\arguments{
\item{x}{numeric vector}

\item{mu}{position of function center}
}
\description{
VBtop function to store VBtop position in a CoreLine@Components slot
}
