% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{WidgetState}
\alias{WidgetState}
\title{WidgetState}
\usage{
WidgetState(widget, value)
}
\arguments{
\item{widget}{the toplevel widget container}

\item{value}{is 'normal' or 'disabled'}
}
\description{
WidgetState sets the state of all the widget childrens
}
