% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{xps.info_GUI}
\alias{xps.info_GUI}
\title{xps.info_GUI Info GUI for XPSSample}
\usage{
xps.info_GUI(xps.sample_name = NULL)
}
\arguments{
\item{xps.sample_name}{name of XPSSample}
}
\description{
xps.info_GUI XPSSample info GUI
}
\author{
Roberto Canteri \email{canteri@fbk.eu}
}
