% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSAnalysisTKGUI.r
\name{XPSAnalysis}
\alias{XPSAnalysis}
\title{XPSAnalysis to analyze XPS-Corelines and XPS spectra in general}
\usage{
XPSAnalysis()
}
\value{
Returns the processed \code{object}.
}
\description{
XPSAnalysis() function is an interactive GUI to add BaseLines
  and Fitting components to a CoreLine needed for the best fit and quantification
}
\examples{
\dontrun{
   XPSAnalysis()
}
}
