% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSAnnotateTKGUI.r
\name{XPSAnnotate}
\alias{XPSAnnotate}
\title{XPSAnnotate is graphic function to add text to lattice plots}
\usage{
XPSAnnotate()
}
\description{
XPSAnnotate() adds text to spectra plotted using (\code{matplot})base function.
}
\examples{
\dontrun{
	XPSAnnotate()
}
}
\seealso{
\link{matplot}, \link{plot}
}
