% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{XPSbaseline,XPSCoreLine-method}
\alias{XPSbaseline,XPSCoreLine-method}
\title{XPSbaseline}
\usage{
\S4method{XPSbaseline}{XPSCoreLine}(
  object,
  bgtype = c("linear", "shirley", "polynomial", "spline"),
  deg = NULL,
  Wgt = NULL,
  splinePoints = list(x = NULL, y = NULL),
  ...
)
}
\arguments{
\item{object}{XPSCoreLine object}

\item{bgtype}{the baseline type. For example linear type \code{"2P.Tougaard"}.}

\item{deg}{degree of the polynomial background}

\item{Wgt}{LinearPolynomial weigth in LPShirley, required parma in 3P. 4P.Tougaard}

\item{splinePoints}{numeric vector containing the points which will be connected by the spline}

\item{...}{other parameters.}
}
\value{
The Object slot \code{Baseline}
}
\description{
method to generate a Baseline for an object of class 'XPSCoreLine'
}
\examples{
\dontrun{
  test[["C1s"]] <- XPSbaseline(test[["C1s"]], "linear")
}
}
