% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{XPScalc}
\alias{XPScalc}
\title{XPScalc function to calculate the element concentrations}
\usage{
XPScalc(object, table = TRUE)
}
\arguments{
\item{object}{XPSCoreLine}

\item{table}{Print table Logic}
}
\value{
An ASCII file for each XPSCoreLine.
}
\description{
XPScalc Function to compute the integral intensity for a given XPS-CoreLine.
 Calculation of Integral Intensity of fitting components of a XPSCoreLine.
}
\seealso{
\link{write.table}, \link{write.csv}, \link{write.csv2}
}
