% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSCompareEngine.r
\name{XPScompEngine}
\alias{XPScompEngine}
\title{XPScompare allows compare Core-Line and their best fits}
\usage{
XPScompEngine(PlotParameters, Plot_Args, SelectedNames, Xlim, Ylim)
}
\arguments{
\item{PlotParameters}{the plot parameters asociated to the XPSOverlayGUI options;}

\item{Plot_Args}{list of plot options;}

\item{SelectedNames}{list containing the XPSSample names and the Corelines to be plotted;}

\item{Xlim}{Xrange of the data to be plotted;}

\item{Ylim}{Yrange of the data to be plotted;}
}
\value{
Returns c(Xlim, Ylim) eventually modified
}
\description{
XPScompEngine is the software powering the XPSCompareGUI.
  XPScompEngine() is called by XPScompareGUI() and not directly accesible.
  This macro compares corelines in multi-panel mode following
  options selected by the user in XPScompareGUI
}
