% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSConstraintsTKGUI.r
\name{XPSConstraints}
\alias{XPSConstraints}
\title{XPSConstraints is a user interface to add a fit constraints}
\usage{
XPSConstraints()
}
\description{
XPSConstraints function is a user interface to simplify
  It adds constraints among fit components defined for a XPSCoreLine object.
  This function is called after the definition of the baseline using 
  (\code{XPSbaseline}) and fitting components with (\code{XPSAddFitComponent}).
}
\examples{
\dontrun{
	XPSConstraints()
}
}
\seealso{
\link{XPSConstrain}, \link{XPSAddFitComponent}, \link{XPSFitAlgorithms}
}
