% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSCoreLineList.r
\name{XPSCoreLineList}
\alias{XPSCoreLineList}
\title{XPSCoreLineList provides the XPS-Sample Core-Lines}
\usage{
XPSCoreLineList()
}
\description{
XPSCoreLineList gives the list of Core-Lines 
  which are contained in the active XPS-Sample. This list
  corresponds to the Core-Lines plotted when double-clicking
  on a XPS-Sample loaded in RxpsG. This function is commonly 
  used by the other RxpsG functions
}
\examples{
\dontrun{
	XPSCoreLineList()
}
}
