% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSElementsTK.r
\name{XPSDefineRSF}
\alias{XPSDefineRSF}
\title{XPSDefineRSF}
\usage{
XPSDefineRSF(Object = NULL, Symbol = NULL)
}
\arguments{
\item{Object}{of class XPSCoreLine, the selected Core.Line}

\item{Symbol}{of class character, the name of the selected Core.Line}
}
\description{
XPSDefineRSF() function select the RSF for the
  selected Core.Line. Given the name of the Core.Line (C1s, Cu2p, Au4f...)
  XPSSetRSF() separate the Element name and the Orbital. Then
  the relative RSF is obtained from tabulated data.
  If more than one value is found a GUI is generated to manually select the RSF
  If no Object is specified the active Core.Line is taken
  If no Symbol is specified the activeSpectName is taken
}
\examples{
  \dontrun{
  XPSDefineRSF(XPSSample[["W4f"]], "W4f")
  XPSDefineRSF(XPSSample[[2]], "C1s")
  XPSDefineRSF()
}
}
