% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSDiffTKGUI.r
\name{XPSDiff}
\alias{XPSDiff}
\title{XPSDiff for spectra differentiation}
\usage{
XPSDiff()
}
\description{
XPSDiff function is used to compute the derivative of the XPSSpectra.
  The user can select the degree of differentiation. Tje result of the differentiation
  is oplotted together with the original data. It is possible to amplify the
  differentiated spectrum to be on the scale of the original spectrum.
}
\examples{
\dontrun{
	XPSDiff()
}
}
