% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSElemTabTKGUI.r
\name{XPSElemTab}
\alias{XPSElemTab}
\title{XPSElemTab}
\usage{
XPSElemTab()
}
\description{
XPSElemTab constructs the CoreLine and Auger Transition tables
  this function helps the identification of chemical elements in Survey spectra.
}
\examples{
\dontrun{
	XPSElemTab()
}
}
