% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSExportAsciiTKGUI.r
\name{XPSExportAscii}
\alias{XPSExportAscii}
\title{XPSExportAscii}
\usage{
XPSExportAscii()
}
\description{
XPSExportAscii GUI for the selection of the more convenient
  style to export single and all the XPSCoreLines in a XPSSample in anASCII 
  text file.
}
\examples{
\dontrun{
 XPSexportASCII()
}
}
