% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{XPSfitAlgorithms}
\alias{XPSfitAlgorithms}
\title{XPSfitAlgorithms}
\usage{
XPSfitAlgorithms()
}
\value{
Returns the selected fit algorithm.
}
\description{
XPSfitAlgorithms groups all the definitions of the fit functions,
  the definition of fit parameters, and the initialization of the parameter
  slots of objects of class 'XPSCoreLine'.
}
\examples{
\dontrun{
XPSfitAlgorithms()
}
}
