% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitInfo.r
\name{XPSFitInfo}
\alias{XPSFitInfo}
\title{XPSFitInfo}
\usage{
XPSFitInfo()
}
\value{
XPSFitInfo returns a table summarizing the area, FWHM, RSF, BE, % for
  each of the fit components if the active CoreLine
}
\description{
XPSFitInfo Provides information about a Core Line fit
  Makes the computation of the integral intensity of each of the fitting components
  of a given coreline. The sum of the component integral intensities =100% (it is the
  best fit integral intensity).
}
\examples{
\dontrun{
	XPSFitInfo()
}
}
