% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitParamTKGUI.r
\name{XPSFitParam}
\alias{XPSFitParam}
\title{XPSFitParam}
\usage{
XPSFitParam(ParamNames, ParamValues)
}
\arguments{
\item{ParamNames}{c("maxiteration", "tolerance", "minfactor") the set of parameter to work with}

\item{ParamValues}{the correspondent values}
}
\description{
XPSFitParam GUI to set the fit param MaXiTERATION, TOLERANCE, MINFACTOR
 when NLSfit error occurs
}
\examples{
\dontrun{
 XPSFitParam(c("maxiteration", "tolerance", "minfactor"), c(1000, 1e-5, 1e-3))
}
}
