% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSImportAsciiTKGUI.r
\name{XPSImport.Ascii}
\alias{XPSImport.Ascii}
\title{XPSImport.Ascii function}
\usage{
XPSImport.Ascii()
}
\description{
Function allowing import of textual (ascii) file data
  Options are available to account for header and data separator
  Options are also provided to store data in an object of class 'XPSSample'
  XPSImport.Ascii saves the imported data in the .GlobalEnv
}
\examples{
\dontrun{
 XPSImport.Ascii()
 activeFName <- get("activeFName", envir=.GlobalEnv)
 XPSSample <- get(activeFName, envir=.GlobalEnv)
}
}
