% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSModFitTKGUI.r
\name{XPSModFit}
\alias{XPSModFit}
\title{XPSmodFit performs the fit of a Core Line}
\usage{
XPSModFit(Object, plt = TRUE, ...)
}
\arguments{
\item{Object}{XPSCoreLine object}

\item{plt}{logical, to enable plot XPSCoreLine and fit. By default is set to TRUE}

\item{\dots}{further parameters to the fitting function}
}
\value{
The Object slot \code{XPSmodFit} will be filled with the result of the
 calculation. All the values of the components will be update and the result
 will be displyed.
}
\description{
Provides a userfriendly interface to select a fitting algorithms to
  fit an XPSCoreline. Fitting algorithms are:
  (I) Classic algorithms: Levenberg-Marquardt, Newton and Port based on the minimization
    of the Squares of the Differences and handled in a more robust way (although slower)
    compared to (\code{nlsLM}) function.
  (II) Conjugate-Gradient Algorithms: General-purpose optimization based on Nelder-Mead,
    quasi-Newton and conjugate-gradient algorithms.
  (III) Pseudo Algorithms: random based algorithm (rather slow) but ensure the convergence.
}
\examples{
\dontrun{
	SampData[["C1s"]] <- XPSmodFit(SampData[["C1s"]])
}
}
\seealso{
\link{nlsLM}
}
