% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSMoveCompTKGUI.r
\name{XPSMoveComp}
\alias{XPSMoveComp}
\title{XPSMoveComp}
\usage{
XPSMoveComp()
}
\description{
XPSMoveComp function to modify component position and intensity in a fit
  this function provides a userfriendly interface change position and intensity of each
  individual fitting component of a selected XPSCoreline. Changes are saved
  in the .GlobalEnv main software memory.
}
\examples{
\dontrun{
	XPSMoveComp()
}
}
