% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSOverlayEngineTK.r
\name{XPSovEngine}
\alias{XPSovEngine}
\title{XPSovEngine}
\usage{
XPSovEngine(PlotParameters, Plot_Args, SelectedNames, Xlim, Ylim)
}
\arguments{
\item{PlotParameters}{the plot parameters asociated to the XPSOverlayGUI options;}

\item{Plot_Args}{list of plot options;}

\item{SelectedNames}{list containing the XPSSample names and the Corelines to be plotted;}

\item{Xlim}{Xrange of the data to be plotted;}

\item{Ylim}{Yrange of the data to be plotted;}
}
\value{
Return c(Xlim, Ylim) eventually modified
}
\description{
XPSOverlay Engine is a macro called by XPSOverlayGUI
   which uses lattice to plot data following the graphical options
}
