% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSOverlayTKGUI.r
\name{XPSOverlay}
\alias{XPSOverlay}
\title{XPSOverlay}
\usage{
XPSOverlay()
}
\description{
function to superpose plots of CoreLine spectra
  Provides a userfriendly interface to select XPS-Corelines to overlay
  and a selection of plotting options for a personalized data representation
  This function is based on the (\code{Lattice}) package.
}
\examples{
\dontrun{
	XPSOverlay()
}
}
