% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{XPSpkgCtrl,XPSCoreLine-method}
\alias{XPSpkgCtrl,XPSCoreLine-method}
\title{XPSpkgCtrl}
\usage{
\S4method{XPSpkgCtrl}{XPSCoreLine}(object)
}
\arguments{
\item{object}{a Core_Line object of class \code{XPSCoreLine}}
}
\description{
method to control the attribute 'package' of an object of class 'XPSCoreLine'
}
\examples{
\dontrun{
 test[["C1s"]] <- XPSpkgCtrl(test[["C1s"]])
}
}
