% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{XPSpkgCtrl,XPSSample-method}
\alias{XPSpkgCtrl,XPSSample-method}
\title{'XPSpkgCtrl'}
\usage{
\S4method{XPSpkgCtrl}{XPSSample}(object)
}
\arguments{
\item{object}{XPSSample}
}
\description{
method to controls the XPSSamples attributes
}
\examples{
\dontrun{
 test <- XPSpkgCtrl(test) #controls the  attributes of the  test XPSSample
}
}
