% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{XPSpkgCtrl}
\alias{XPSpkgCtrl}
\title{XPSpkgCtrl}
\usage{
XPSpkgCtrl(object)
}
\arguments{
\item{object}{XPSCoreLine object}
}
\value{
'XPSpkgCtrl' returns an object of class 'XPSCoreLine' with attribute '.GlobalEnv'
}
\description{
XPSpkgCtrl Function to control the attributes of an XPSCoreLine.
}
