% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSPreferencesTKGUI.r
\name{XPSPreferences}
\alias{XPSPreferences}
\title{XPSPreferences}
\usage{
XPSPreferences()
}
\description{
XPSPreferences function allows selection of preferences as
  - plot colors, the default set of line patterns and set of symbols
  - the character style used in RxpsG outputs
  - the XPS excitatio source (Al, Mg)
  - the default working directory
  - the default operating system
Preferences are stored in the XPSSettings.ini file saved in the ...Library/RxpsG/extdata.
}
\examples{
\dontrun{
	XPSPreferences()
}
}
