% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSProcessCoreLineTKGUI.r
\name{XPSProcessCoreLine}
\alias{XPSProcessCoreLine}
\title{XPSProcessCoreLine}
\usage{
XPSProcessCoreLine()
}
\description{
XPSProcessCoreLine function adds a CoreLine (with fit if existing)
  to an object of class XPSSample, or just a BaseLine and Fit to an existing CoreLine
  (class XPSCoreLine) or remove a CoreLine from an object of class XPSSample
  XPSProcessCoreLine may be used to perform some simple math operations on CoreLines
  (class XPSCoreLine) like addition, subtraction, multiplication by a constant value
   normalization, differentiation, combination of two CoreLines.
}
\examples{
\dontrun{
	XPSProcessCoreLine()
}
}
