% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{XPSquantify,XPSSample-method}
\alias{XPSquantify,XPSSample-method}
\title{'XPSquantify'}
\usage{
\S4method{XPSquantify}{XPSSample}(object, without = NULL, verbose = TRUE)
}
\arguments{
\item{object}{XPSSample}

\item{without}{character vector of coreline names which will be not used in the quantification}

\item{verbose}{if TRUE prints the quantification table}
}
\description{
method compute the element quantification of objects of class 'XPSSample'
}
\examples{
\dontrun{
 QuantData <- XPSquantify(test, without=NULL, verbose=TRUE ) #computes the  quantification for the  test XPSSample
}
}
