% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{XPSquantify}
\alias{XPSquantify}
\title{'XPSquantify'}
\usage{
XPSquantify(object, without = NULL, verbose = TRUE)
}
\arguments{
\item{object}{XPSSample}

\item{without}{character vector of coreline names which will be not used in the quantification}

\item{verbose}{if TRUE prints the quantification table}
}
\description{
compute the element quantification for the selected XPSSample
}
