% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSRead.OldScienta.r
\name{XPSRead.Oldscienta}
\alias{XPSRead.Oldscienta}
\title{XPSRead.Oldscienta}
\usage{
XPSRead.Oldscienta(
  filename = NULL,
  project = "",
  experiment = "",
  out = c("list")
)
}
\arguments{
\item{filename}{file name or list of file names}

\item{project}{comment}

\item{experiment}{comment}

\item{out}{output class type default list()}
}
\value{
XPSRead.Oldscienta returns and object of class XPSSample
}
\description{
XPSRead.Oldscienta function is used to read XPS data
  in the old-DOS-Scienta format
  Data are saved in the following folder which must be present
   /Analysis
   /Regions
   /Results
   /Specific
  Selection of the DataFile in the /Analysis folder activates reading
  data in all the folders.
}
\examples{
\dontrun{
 pk1 <- read.xps("c:/R/Analysis/OldScientaFiles/Analysis/DLC.48")
}
}
