% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{XPSremove,XPSCoreLine-method}
\alias{XPSremove,XPSCoreLine-method}
\title{XPSremove}
\usage{
\S4method{XPSremove}{XPSCoreLine}(
  object,
  what = c("all", "fit", "components", "baseline", "regionToFit"),
  number
)
}
\arguments{
\item{object}{XPSCoreLine}

\item{what}{one of "all", "fit", "components", "baseline", "regionToFit"}

\item{number}{in case of \code{what='components'}: the component number will
be reset,\cr if it is missing then all the components will be removed.}
}
\description{
method to erase the content of selected slots an object of class 'XPSCoreLine'
}
\examples{
\dontrun{
 test[["C1s"]] <- XPSremove(test[["C1s"]], "baseline")
 test[["C1s"]] <- XPSremove(test[["C1s"]], "all") # this reset everything
}
}
