% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{XPSremove}
\alias{XPSremove}
\title{XPSremove Rests selected processing elements from XPSCoreLine}
\usage{
XPSremove(
  object,
  what = c("all", "fit", "components", "baseline", "regionToFit"),
  number
)
}
\arguments{
\item{object}{XPSCoreLine}

\item{what}{one of "all", "fit", "components", "baseline", "regionToFit"}

\item{number}{in case of \code{what='components'}: the component number will
be reset,\cr if it is missing then all the components will be removed.}
}
\value{
'XPSremove' returns the object XPSCoreLine with erased slots.
}
\description{
The function is designed to erase the content of selected slots: \cr \code{"all"} =
  it resets all the \code{slots}. This is the default value if \code{what} is
  missing. See examples. \cr \code{"fit"} = it resets the \code{Fit slot}. \cr
  \code{"components"} = it resets the \code{Components slot} indicated by the
  \code{number}. If \code{number} is missing then all the components will be
  reset.\cr \code{"baseline"} = it resets the \code{Baseline slot}. \cr
  \code{"regionToFit"} = it removes only and any link. \cr
}
\seealso{
\linkS4class{XPSCoreLine}
}
