% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSResetAnalysisTKGUI.r
\name{XPSResetAnalysis}
\alias{XPSResetAnalysis}
\title{XPSResetAnalysis}
\usage{
XPSResetAnalysis()
}
\description{
XPSResetAnalysis function resets the analysis performed
  Function to delete individual elements of a Coreline analysis:
  resets only the Best Fit;
  resets an single or all the Fitting Components;
  resets the BaseLine (i.e. eliminates alsio the Fit);
  resets ALL alle the Coreline analysis.
}
\examples{
\dontrun{
	XPSResetAnalysis()
}
}
