% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSetCoreLineTKGUI.r
\name{XPSSetCoreLine}
\alias{XPSSetCoreLine}
\title{XPSSetCoreLine}
\usage{
XPSSetCoreLine()
}
\description{
XPSSetCoreLine function to select a Core Line
  The list of corelines of a give XPS-Sample are presented for selection
}
\examples{
\dontrun{
XPSSetCoreLine()
}
}
