% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSetGraphDevTKGUI.r
\name{XPSSetGraphDev}
\alias{XPSSetGraphDev}
\title{XPSSetGraphDev}
\usage{
XPSSetGraphDev()
}
\description{
function to select the kind of graphic device compatible
  with the operating system in use.
  Also a list of graphic formats is provided to save the content 
  of the current graphic device.
}
\examples{
\dontrun{
	XPSSetGraphDev()
}
}
