% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{XPSsetRegionToFit,XPSCoreLine-method}
\alias{XPSsetRegionToFit,XPSCoreLine-method}
\title{getMaxOfComponents}
\usage{
\S4method{XPSsetRegionToFit}{XPSCoreLine}(object, limits, ...)
}
\arguments{
\item{object}{XPSCoreLine object}

\item{limits}{list with x,y values to limit the region.}

\item{...}{further parameters to the XPSsetRegionToFit function}
}
\description{
method to define the maximum of Fit Components
}
\examples{
\dontrun{
 test[["C1s"]] <- XPSsetRegionToFit(test[["C1s"]], limits=pos)   
}                                                                
}
