% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{XPSsetRegionToFit}
\alias{XPSsetRegionToFit}
\title{XPSsetRegionToFit Definition of the Region to Fit}
\usage{
XPSsetRegionToFit(object, limits, ...)
}
\arguments{
\item{object}{XPSCoreLine object}

\item{limits}{list with x,y values to limit the region.}

\item{...}{further parameters to the XPSsetRegionToFit function}
}
\value{
returns a portion of the CoreLine for a backgrouns subtraction
}
\description{
Definition of the portion (usually on the x-axis) of the original curve for
  further processing. XPSsetRegionToFit limits usually will be set with the cursor on the 
  curve plot where a \code{Baseline} is needed. Then the \code{RegionToFit} is
  defined as a list of (x,y) data equal to the selected portion of the curve.
}
\seealso{
\link{XPSbaseline}
}
