% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{XPSsetRSF}
\alias{XPSsetRSF}
\title{XPSsetRSF set the value of the RSF}
\usage{
XPSsetRSF(object, rsf = NULL)
}
\arguments{
\item{object}{XPSCoreLine}

\item{rsf}{RSF value}
}
\description{
XPSsetRSF Set the RSF value of XPSCoreLine for quantification.
}
