% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSpectList.r
\name{XPSSpectList}
\alias{XPSSpectList}
\title{XPSSpectList}
\usage{
XPSSpectList(SelectedFName, noIdx = FALSE)
}
\arguments{
\item{SelectedFName}{the name (string) of the selected XPSSample}

\item{noIdx}{logical if FALSE a numerical index is added to any CoreLine name}
}
\description{
XPSSpectList shows the list of spctra related to
  object of class XPSSample. The spectra are identified by their
  an index or by the CoreLine name
}
\examples{
\dontrun{
	XPSSampleInfo(SelectedFName, noIdx=FALSE)
}
}
