% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSpectNameChangeTKGUI.r
\name{XPSSpectNameChange}
\alias{XPSSpectNameChange}
\title{XPSSpectNameChange}
\usage{
XPSSpectNameChange()
}
\description{
XPSSpectNameChange is a function to change the name of
  objects of class XPSSample and/or the name associated to CoreLines
  of class XPSCoreLine
}
\examples{
\dontrun{
	XPSSpectNameChange()
}
}
