% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSplitTKGUI.r
\name{XPSSplit}
\alias{XPSSplit}
\title{XPSSplit}
\usage{
XPSSplit()
}
\description{
XPSSplit to select split multiple acquisitions performed
  Acquisitions on multiple samples may be included in a single .PXT file
  This function allows splitting spectra corresponding to each sample in
  individual files.
}
\examples{
\dontrun{
	XPSSplit()
}
}
