% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSurElmtIdentifyTKGUI.r
\name{XPSSurveyElementIdentify}
\alias{XPSSurveyElementIdentify}
\title{XPSSurveyElementIdentify}
\usage{
XPSSurveyElementIdentify()
}
\description{
function to Identify Elements present in a Survey spectrum.
}
\examples{
\dontrun{
XPSSurveyElementIdentify()
}
}
