% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSwitchBE.KE.TKGUI.r
\name{XPSSwitch.BE.KE}
\alias{XPSSwitch.BE.KE}
\title{XPSSwitch.BE.KE}
\usage{
XPSSwitch.BE.KE()
}
\description{
XPSSwitch.BE.KE function to set the energy scale in an object of class XPSSample
  The energy scale in a XPSSample may be set as Binding Energy or Kinetic Energy
  This function converts Binding energy into Kinetic and viceversa. The selected energy scale
  for a given CoreLine will be used  to plot the correspondent spectral data.
}
\examples{
\dontrun{
	XPSSwitch.BE.KE()
}
}
