% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{XPSTable}
\alias{XPSTable}
\title{XPSTable}
\usage{
XPSTable(parent, items, NRows = 0, ColNames, Width)
}
\arguments{
\item{parent}{the widget container}

\item{items}{the items used to generate the Table}

\item{NRows}{the number of Table rows}

\item{ColNames}{the name of the Table columns}

\item{Width}{the width of the Table columns}
}
\description{
TableList generates a table-like widget with the specified items
}
