% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSTwoScalePlotTKGUI.r
\name{XPSTwoScalePlot}
\alias{XPSTwoScalePlot}
\title{XPSTwoScalePlot}
\usage{
XPSTwoScalePlot()
}
\description{
XPSTwoScalePlot performs plots overlapping two different
  object with different Y scales
  Provides a userfriendly interface to select Objects to superpose
  and a selection of major plotting options for a personalized data representation
  Data are represented using different Left and Right Y scales
}
\examples{
\dontrun{
	XPSTwoScalePlot()
}
}
