% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSVmsCorrTKGUI.r
\name{XPSVmsCorr}
\alias{XPSVmsCorr}
\title{XPSVmsCorr}
\usage{
XPSVmsCorr()
}
\description{
XPSVmsCorr function corrects Vamas_type data for the analyzer transmission
  This routine divides raw data intensity for the value of the analyzer transmission
  The routine applies the correction to the selected coreline or to all the XPSSample
  spectra, baselines and fit components when present.
  The transmission function must be provided by the XPS-Instrument manufacturer
}
\examples{
\dontrun{
	XPSVmsCorr()
}
}
