% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitCompClass.r
\name{Ycomponent}
\alias{Ycomponent}
\title{Ycomponent}
\usage{
Ycomponent(object, x, y)
}
\arguments{
\item{object}{a fit component object of the type CL@Components[[j]] where
CL is of class 'XPSCoreLine', j indicates the jth fit component}

\item{x}{numeric vector of the type CL@RegionToFit$x where
CL is of class 'XPSCoreLine'}

\item{y}{numeric vector of the type CL@Baseline$y where
CL is of class 'XPSCoreLine'}
}
\value{
'Ycomponent' returns Baseline subtracted vales of the fit component
}
\description{
S4method 'Ycomponent' method for generate the Y values of the
  fit component using the chosen fitting function and the associated fit parameters
}
