% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{yscale.components.logpower10ticks}
\alias{yscale.components.logpower10ticks}
\title{yscale.components.logpower10ticks}
\usage{
yscale.components.logpower10ticks(lim, logsc = FALSE, at = NULL, ...)
}
\arguments{
\item{lim}{lim.}

\item{logsc}{logsc.}

\item{at}{at.}

\item{...}{other plot parameters}
}
\description{
yscale.components.logpower10ticks plots log ticks.
}
