% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{yscale.components.power10ticks}
\alias{yscale.components.power10ticks}
\title{yscale.components.power10ticks}
\usage{
yscale.components.power10ticks(
  lim,
  packet.number = 0,
  packet.list = NULL,
  right = TRUE,
  ...
)
}
\arguments{
\item{lim}{lim.}

\item{packet.number}{packet.number.}

\item{packet.list}{packet.list.}

\item{right}{right.}

\item{...}{other plot parameters}
}
\description{
yscale.components.power10ticks function to label the ticks plot with exponential format.
}
