% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RXPSTKGUI.r, R/RxpsG-package.r
\docType{package}
\name{RxpsG}
\alias{RxpsG}
\alias{xps}
\alias{RxpsG-package}
\title{'RxpsG' is the main interface of the Package RxpsG}
\usage{
xps()
}
\description{
xps() activates the RxpsG front-end to use
  all the functions to load raw data perform the spectral
  analysis and plot raw and analyzed spectra

Processing tools for X ray Photoelectron Spectroscopy Data.
  The package implements classes and methods to read, visualize and manipulate XPS spectra files.
  The spectra can be obtained from .pxt and .vms data format from different instruments.'
  More generally, any data that is recorded as a list of (x,y) is suitable.
}
\examples{
 \dontrun{
 xps()
}
}
\keyword{internal}
